Update behavior of WarnHistoryCommand and fix overflow bug in parseStringtoMillisec. WarnHistoryCommand now shows warnings and active bans for players with permission gesuit.bans.command.warn Fix overflow bug in parseStringtoMillisec

+++ src/main/java/net/cubespace/geSuitBans/utils/TimeParser.java
@@ -52,7 +52,7 @@ public class TimeParser
     }
 
     public static long parseStringtoMillisec(String string) {
-        int total = parseStringToSecs(string);
+        long total = parseStringToSecs(string);
         return total * 1000;
     }
 }
